Tratamento de erro, correcao do backend, adicao procura de projetos

+++ devfy/src/main/java/br/com/devfy/devfy/controller/DesenvolvedorController.java
@@ -1,5 +1,6 @@
 package br.com.devfy.devfy.controller;
 
+import br.com.devfy.devfy.helper.ListaObj;
 import br.com.devfy.devfy.model.Desenvolvedor;
 import br.com.devfy.devfy.repository.DesenvolvedorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,8 +19,8 @@ public class DesenvolvedorController {
 
     @GetMapping
     public ResponseEntity exibir() {
-        List<Desenvolvedor> desenvolvedores = repository.findAll();
-        if (desenvolvedores.isEmpty()) {
+        ListaObj<Desenvolvedor> desenvolvedores = (ListaObj<Desenvolvedor>) repository.findAll();
+        if (desenvolvedores.getTamanho()==0) {
             return ResponseEntity.status(204).build();
         }
         return ResponseEntity.status(200).body(desenvolvedores);
@@ -31,17 +32,23 @@ public class DesenvolvedorController {
             @PathVariable int id,
             @RequestBody Desenvolvedor desenvolvedor
     ) {
-        desenvolvedor.setId(id);
-        repository.save(desenvolvedor);
-        return ResponseEntity.status(200).build();
+        if (repository.existsById(id)) {
+            desenvolvedor.setId(id);
+            repository.save(desenvolvedor);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
     }
 
     @DeleteMapping("/deletar/{id}")
     public ResponseEntity deletar(
             @PathVariable int id) {
-        repository.deleteById(id);
-        return ResponseEntity.status(200).build();
-    }
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
+        }
 
     @PostMapping("/cadastrar")
     public ResponseEntity adicionar(
@@ -51,20 +58,28 @@ public class DesenvolvedorController {
         return ResponseEntity.status(201).build();
     }
 
-    @GetMapping("/login/{id}")
-    public ResponseEntity login(@PathVariable int id) {
-        Desenvolvedor desenvolvedor = repository.getById(id);
-        desenvolvedor.login();
-        repository.save(desenvolvedor);
-        return ResponseEntity.status(200).build();
+    @GetMapping("/login")
+    public ResponseEntity login(@RequestBody Desenvolvedor desenvolvedor) {
+        List<Desenvolvedor> desenvolvedores = repository.findAll();
+        for (Desenvolvedor desenvolvedor1 : desenvolvedores){
+            if (desenvolvedor1.login(desenvolvedor)){
+                desenvolvedor1.setIsAutenticado(true);
+                repository.save(desenvolvedor1);
+                return ResponseEntity.status(200).build();
+            }
+        }
+        return ResponseEntity.status(404).build();
     }
 
     @GetMapping("/logoff/{id}")
     public ResponseEntity logoff(@PathVariable int id) {
-        Desenvolvedor desenvolvedor = repository.getById(id);
-        desenvolvedor.logoff();
-        repository.save(desenvolvedor);
-        return ResponseEntity.status(200).build();
+        if (repository.existsById(id)) {
+            Desenvolvedor desenvolvedor = repository.getById(id);
+            desenvolvedor.logoff();
+            repository.save(desenvolvedor);
+            return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
     }
 
 }
