+++ devfy/src/main/java/br/com/devfy/devfy/controller/ProjetoController.java
@@ -7,6 +7,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.ListIterator;
 
+import br.com.devfy.devfy.model.Desenvolvedor;
 import br.com.devfy.devfy.model.Projeto;
 import br.com.devfy.devfy.repository.ProjetoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,9 +24,9 @@ public class ProjetoController {
 
     @GetMapping
     public ResponseEntity exibir() {
-        List<Projeto> projetos = repository.findAll();
+        ListaObj<Projeto> projetos = (ListaObj<Projeto>) repository.findAll();
 
-        if (projetos.isEmpty()) {
+        if (projetos.getTamanho()==0) {
             return ResponseEntity.status(204).build();
         }
         return ResponseEntity.status(200).body(projetos);
@@ -37,16 +38,22 @@ public class ProjetoController {
             @PathVariable int id,
             @RequestBody Projeto projeto
     ) {
+        if (repository.existsById(id)){
         projeto.setId(id);
         repository.save(projeto);
         return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
     }
 
     @DeleteMapping("/deletar/{id}")
     public ResponseEntity deletar(
             @PathVariable int id) {
+        if (repository.existsById(id)){
         repository.deleteById(id);
         return ResponseEntity.status(200).build();
+        }
+        return ResponseEntity.status(404).build();
     }
 
     @PostMapping("/cadastrar")
@@ -56,5 +63,28 @@ public class ProjetoController {
         repository.save(projeto);
         return ResponseEntity.status(201).build();
     }
+
+    @GetMapping("/procurar/{nomeProjeto}")
+    public ResponseEntity getProjetoByNome(@PathVariable String nomeProjeto){
+        List<Projeto> projetos = repository.findAll();
+        for (Projeto projeto : projetos){
+            if (nomeProjeto.equals(projeto.getTitulo())){
+                return ResponseEntity.status(200).body(projeto);
+            }
+        }
+        return ResponseEntity.status(404).build();
+    }
+
+    @GetMapping("/associarDev/{idDev}/{idProj}")
+    public ResponseEntity associarDev(@PathVariable int idDev, @PathVariable int idProj){
+            if (repository.existsById(idProj)){
+                Projeto projeto = repository.getById(idProj);
+                projeto.setFkDev(idDev);
+                repository.save(projeto);
+                return ResponseEntity.status(200).build();
+            }
+            return ResponseEntity.status(404).build();
+    }
+
 
 }
